+++ src/counters/atomic_counter.rs
@@ -27,36 +27,30 @@ mod tests {
     use counters::Counter;
     use std::thread;
     use std::sync::Arc;
+    use test::Bencher;
+    use counters::counter::test_utils::*;
 
-    fn spawn_incr(sc: Arc<AtomicCounter>, n: isize) -> thread::JoinHandle<()> {
-        thread::spawn(move || {
-            for _ in 0..n { sc.inc(1) }
-        })
-    }
+    fn ctor() -> AtomicCounter { AtomicCounter::new() }
 
     #[test]
     fn test_atomic_counter() {
-        let c = Arc::new(AtomicCounter::new());
-
         let thread_count = 16;
         let iter_count = 1000000;
 
+        let c = test_counter(&ctor, iter_count, thread_count);
 
-         let mut children = vec![];
-
-         for _ in 0..thread_count {
-             children.push(spawn_incr(c.clone(), iter_count));
-         }
-
-         for child in children {
-             let res = child.join();
-             assert!(res.is_ok());
-         }
-
-        assert!(c.snapshot() as isize == thread_count * iter_count);
+        assert!(c.snapshot() == thread_count * iter_count);
 
         c.clear();
 
-        assert!(c.snapshot() as isize == 0);
+        assert!(c.snapshot() == 0);
+    }
+
+    #[bench]
+    fn bench_atomic_counter(b: &mut Bencher) {
+        let thread_count = 16;
+        let iter_count = 10000;
+
+        b.iter(|| { test_counter(&ctor, iter_count, thread_count) })
     }
 }
